# Improve LoRA Perf by Deprecating FlashInfer and Eliminating Redundant Tensor Ops