<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation
1. Deprecate FlashInfer LoRA backend: #7809
2. Optimize prepare_lora_batch to optimize latency as currently there are lots of redundant mem ops, many of which for backward compatibility with FlashInfer: #6961, #8712 

## Modifications

1. Deprecate FlashInfer LoRA backend
2. Refactor weight shape & prepare_lora_batch to reduce redundant memory ops.
  - Store lora B weight as (rank, q + k + v) following triton kernel expectation and industry convention.
  - Eliminate all reshape / copy ops in set_lora
  - Move set_lora() calls from prepare_lora_batch to server initialization phase to avoid overheads per batch
3. ~Fix bug #8939 (TODO: rebase once #8987 is merged)~


## Accuracy Test

Covered by CI

## Benchmark & Profiling
### Benchmark
```
python3 -m sglang.launch_server --model-path meta-llama/Llama-3.1-8B-Instruct  --disable-radix-cache --lora-paths lora=algoprog/fact-generation-llama-3.1-8b-instruct-lora
python3 -m sglang.bench_serving --backend sglang  --num-prompt 480 --request-rate 8 --lora-name lora
```

| Metrics | Before | After |
|---------|--------|--------|
| ITL P95 | 54.69ms | 38.40 **(-29.8%)** |
| ITL P50 | 23.35ms | 18.30 **(-21.6%)** |
| E2E Latency P50 | 2996.55ms | 2323.63ms **(-22%)** |
| Full Result | <img width="428" height="579" alt="image" src="https://github.com/user-attachments/assets/19ff56a1-aa48-435b-8964-432a06e90aec" /> | <img width="426" height="572" alt="image" src="https://github.com/user-attachments/assets/13a27608-1534-4727-ab9e-88260ab1ee6c" /> | 

### Profile: Prepare_lora_batch (3.162ms -> 162 us)
- Before 
<img width="955" height="305" alt="image" src="https://github.com/user-attachments/assets/0b5453a5-72a3-48fc-a1f6-c3d802a1d7a6" />

- After
<img width="1029" height="303" alt="image" src="https://github.com/user-attachments/assets/727c4480-f914-41b4-b746-b3274941a24d" />



## Checklist

- [ ] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.